project/requirements.txt
torch
torchvision
torchaudio
fastapi
uvicorn
python-multipart
numpy
scikit-learn
matplotlib
pillow
tqdm
requests
